INTPYTHON-771 Factor shared test serializers into serializers.py
Move CitySerializer, CountrySerializer, StatusTagSerializer,
ContinentSerializer, and PetOwnerSerializer into a shared
tests/rest_framework_/serializers.py so each test file imports
rather than redefines them.
