+++ tests/rest_framework_/test_polymorphic_embedded_model_serializer_integration.py
@@ -7,15 +7,8 @@ except ImportError:
 
 from django.test import TestCase
 
-from django_mongodb_backend.rest_framework import MongoModelSerializer
-
 from .models import Cat, Dog, PetOwner
-
-
-class PetOwnerSerializer(MongoModelSerializer):
-    class Meta:
-        model = PetOwner
-        fields = "__all__"
+from .serializers import PetOwnerSerializer
 
 
 class PolymorphicEmbeddedModelSerializerReadTests(TestCase):
